// src/generated/models/UserProfileDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { SchoolInformationView } from './SchoolInformationView';
export type UserProfileDTO = {
    studentId?: string;
    email?: string;
    phoneNumber?: string;
    institutionId?: string;
    firstName?: string;
    lastName?: string;
    otherName?: string;
    stateOfOrigin?: string;
    profileImage?: string;
    username?: string;
    gender?: string;
    nin?: string;
    isNinVerified?: boolean;
    dateOfBirth?: string;
    schoolInformationView?: SchoolInformationView;
    referralCode?: string;
};